Start the MDS calendar weeks on Mondays
I'll admit this is mostly because I get myself confused when I look at the MDS calendar and if you think students will be confused if we switch, then let's keep it as is. In an attempt to argue my personal preference with from a more objective angle, Mon is ISO 8601, which is what we promote for date formats otherwise (but I agree is not as important here).